代码提交-mbg 实体类mapper

+++ fairy-cloud-product/src/main/java/com/fairy/cloud/product/model/FlashPromotionSessionExt.java
@@ -10,5 +10,5 @@ public class FlashPromotionSessionExt extends SmsFlashPromotionSession {
     @ApiModelProperty(value = "活动状态#根据时间进行判断",notes = "0->进行中,1->即将开始,2->已结束")
     private Integer sessionStatus;
     @ApiModelProperty(value = "当前秒杀活动ID")
-    private Long flashPromotionId;
+    private Integer flashPromotionId;
 }
